Add super fast version for swift v1.2
Unpacking string into an array of UInt8 to work around swift string
access limitations.  Accessing string at index n and calculating the
length of string is an O(n) operation. This is due to the fact that
unicode UTF strings are encoded in variable length. One way to work around
this limitation is to unpack the string into an array of CodeUnit. Passing
the UInt8 array everywhere instead of the string solves the performance
issue.  We went from 64 minute execution time to 45 second execution
time with this change.